+++ server/src/main/java/net/oneandone/stool/server/Type.java
@@ -29,12 +29,12 @@ public class Type {
     public static final String VALUE_REPLICAS = "replicas";
     public static final String VALUE_IMAGE = "image";
 
-    public static final String VALUE_NOTIFY = "stageNotify";
-    public static final String VALUE_EXPIRE = "stageExpire";
-    public static final String VALUE_COMMENT = "stageComment";
+    public static final String VALUE_CONTACT = "metadataContact";
+    public static final String VALUE_EXPIRE = "metadataExpire";
+    public static final String VALUE_COMMENT = "metadataComment";
 
     public static final String[] STAGE_VALUES = {
-            VALUE_COMMENT, VALUE_EXPIRE, VALUE_NOTIFY
+            VALUE_COMMENT, VALUE_EXPIRE, VALUE_CONTACT
     };
 
     public static final String MAIN_CONTAINER = "main"; // TODO: needed for monitoring stats ...